Reworked Navigator class to take a Robot as a parameter in the constructor. This fixes the recursive constructor call problem, even though it does violate the way that Robot is normally grabbed. I've tested the code and everything looks to be fine, but keep a look out for errors stemming from this.
Small changes to GoalieNavigator and IntArray because of changes to Navigator; basically still the same code.

Added "movedir" as a bluetooth command, usage is "movedir 0" to move toward heading 0. All headings should work, however I think that Jeremy and Tim won't get the exact same direction for the same value. This needs to be fixed so that all the code will run right with all the robots.